Moving test utilities to internal package
